refactor: remove random default value
