+++ dsl/src/commonMain/kotlin/it/unibo/collektive/ID.kt
@@ -1,7 +1,5 @@
 package it.unibo.collektive
 
-import kotlin.random.Random
-
 /**
  * A node identifier.
  */
@@ -10,4 +8,4 @@ interface ID
 /**
  * A node [id] represented by an [Int].
  */
-data class IntId(val id: Int = Random.nextInt(Int.MIN_VALUE, Int.MAX_VALUE)) : ID
+data class IntId(val id: Int) : ID
