+++ dsl/src/commonMain/kotlin/it/unibo/collektive/aggregate/AggregateContext.kt
@@ -1,11 +1,14 @@
-package it.unibo.collektive
+package it.unibo.collektive.aggregate
 
+import it.unibo.collektive.ID
 import it.unibo.collektive.field.Field
 import it.unibo.collektive.messages.AnisotropicMessage
 import it.unibo.collektive.messages.InboundMessage
 import it.unibo.collektive.messages.IsotropicMessage
 import it.unibo.collektive.messages.OutboundMessage
 import it.unibo.collektive.networking.Network
+import it.unibo.collektive.runUntil
+import it.unibo.collektive.singleCycle
 import it.unibo.collektive.stack.Path
 import it.unibo.collektive.stack.Stack
 import it.unibo.collektive.state.State
