+++ dsl/src/commonMain/kotlin/it/unibo/collektive/aggregate/AggregateResult.kt
@@ -1,5 +1,6 @@
-package it.unibo.collektive
+package it.unibo.collektive.aggregate
 
+import it.unibo.collektive.ID
 import it.unibo.collektive.messages.OutboundMessage
 import it.unibo.collektive.state.State
