+++ dsl/src/commonMain/kotlin/it/unibo/collektive/aggregate/ops/AggregateOps.kt
@@ -1,4 +1,4 @@
-package it.unibo.collektive
+package it.unibo.collektive.aggregate
 
 import it.unibo.collektive.field.Field
