+++ dsl/src/commonMain/kotlin/it/unibo/collektive/aggregate/ops/SharingContext.kt
@@ -1,4 +1,4 @@
-package it.unibo.collektive
+package it.unibo.collektive.aggregate
 
 /**
  * The lambda context passed to the [share] function.
