+++ dsl/src/commonTest/kotlin/it/unibo/collektive/aggregate/ExchangeTest.kt
@@ -4,7 +4,6 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import it.unibo.collektive.IntId
-import it.unibo.collektive.aggregate
 import it.unibo.collektive.field.Field
 import it.unibo.collektive.messages.AnisotropicMessage
 import it.unibo.collektive.messages.IsotropicMessage
