+++ dsl/src/commonTest/kotlin/it/unibo/collektive/aggregate/SharingTest.kt
@@ -4,13 +4,11 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import it.unibo.collektive.IntId
 import it.unibo.collektive.SharingContext
-import it.unibo.collektive.aggregate
 import it.unibo.collektive.field.Field
 import it.unibo.collektive.field.max
 import it.unibo.collektive.field.min
 import it.unibo.collektive.network.NetworkImplTest
 import it.unibo.collektive.network.NetworkManager
-import it.unibo.collektive.share
 
 class SharingTest : StringSpec({
     // device ids
