+++ dsl/src/commonTest/kotlin/it/unibo/collektive/alignment/BranchAlignment.kt
@@ -4,8 +4,8 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.collections.shouldHaveSize
 import it.unibo.collektive.IntId
-import it.unibo.collektive.aggregate
-import it.unibo.collektive.neighbouring
+import it.unibo.collektive.aggregate.aggregate
+import it.unibo.collektive.aggregate.neighbouring
 import it.unibo.collektive.stack.Path
 import it.unibo.collektive.utils.getPaths
