+++ dsl/src/commonTest/kotlin/it/unibo/collektive/alignment/TestAlignment.kt
@@ -4,9 +4,9 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.collections.shouldContainAll
 import io.kotest.matchers.shouldBe
 import it.unibo.collektive.IntId
-import it.unibo.collektive.aggregate
-import it.unibo.collektive.neighbouring
-import it.unibo.collektive.share
+import it.unibo.collektive.aggregate.aggregate
+import it.unibo.collektive.aggregate.neighbouring
+import it.unibo.collektive.aggregate.share
 import it.unibo.collektive.stack.Path
 import it.unibo.collektive.utils.getPaths
