+++ dsl/src/commonTest/kotlin/it/unibo/collektive/field/FieldManipulationTest.kt
@@ -4,8 +4,8 @@ import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import it.unibo.collektive.IntId
-import it.unibo.collektive.aggregate
-import it.unibo.collektive.neighbouring
+import it.unibo.collektive.aggregate.aggregate
+import it.unibo.collektive.aggregate.neighbouring
 import it.unibo.collektive.network.NetworkImplTest
 import it.unibo.collektive.network.NetworkManager
